requirements.txt
asgiref==3.6.0
Django==3.2.19
djangorestframework==3.14.0
python-dotenv==0.21.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
